botport/pyproject.toml
[project]
name = "botport"
version = "0.1.0"
description = "Agent-to-agent task routing hub for Captain Claw instances"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [
    { name = "Stevica Kuharski", email = "kstevica@gmail.com" }
]
keywords = ["ai", "agent", "routing", "multi-agent", "a2a"]

dependencies = [
    "aiohttp>=3.9.0",
    "aiosqlite>=0.19.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "pyyaml>=6.0.0",
    "structlog>=23.0.0",
    "httpx>=0.25.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]

[project.scripts]
botport = "botport.server:main"

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.setuptools.packages.find]
include = ["botport*"]

[tool.setuptools.package-data]
botport = ["dashboard/static/**/*"]
